+++ crates/noise-functions-config/src/lib.rs
@@ -11,10 +11,7 @@
 #![cfg_attr(feature = "nightly-simd", feature(portable_simd))]
 #![cfg_attr(docsrs, feature(doc_auto_cfg, doc_cfg_hide), doc(cfg_hide(no_global_oom_handling, feature = "nightly-const-fn-float")))]
 
-use noise_functions::{
-    open_simplex::{Improve3, Improve3Xy, Improve3Xz},
-    *,
-};
+use noise_functions::*;
 
 #[cfg(feature = "nightly-simd")]
 use core::simd::*;
@@ -104,7 +101,6 @@ simple_enum! {
 simple_enum! {
     enum Improve {
         None,
-        Default,
         Xy,
         Xz,
     }
@@ -144,12 +140,6 @@ macro_rules! make {
             noise: Seeded { seed: $self.seed, noise: $noise },
         })
     };
-    ($self:ident, $noise:expr, $improve:ident) => {
-        Box::new(Frequency {
-            frequency: $self.frequency,
-            noise: $improve(Seeded { seed: $self.seed, noise: $noise }),
-        })
-    };
 }
 
 macro_rules! make_fbm {
@@ -161,9 +151,6 @@ macro_rules! make_fbm {
                 .seed($self.seed)
                 .frequency($self.frequency),
         )
-    };
-    ($self:ident, $noise:expr, $improve:ident) => {
-        Box::new($improve($noise.fbm($self.octaves, $self.gain, $self.lacunarity).weighted($self.weighted_strength).seed($self.seed)).frequency($self.frequency))
     };
 }
 
@@ -177,9 +164,6 @@ macro_rules! make_ridged {
                 .frequency($self.frequency),
         )
     };
-    ($self:ident, $noise:expr, $improve:ident) => {
-        Box::new($improve($noise.ridged($self.octaves, $self.gain, $self.lacunarity).weighted($self.weighted_strength).seed($self.seed)).frequency($self.frequency))
-    };
 }
 
 macro_rules! make_ping_pong {
@@ -192,67 +176,9 @@ macro_rules! make_ping_pong {
                 .frequency($self.frequency),
         )
     };
-    ($self:ident, $noise:expr, $improve:ident) => {
-        Box::new(
-            $improve(
-                $noise
-                    .ping_pong($self.octaves, $self.gain, $self.lacunarity, $self.ping_pong_strength)
-                    .weighted($self.weighted_strength)
-                    .seed($self.seed),
-            )
-            .frequency($self.frequency),
-        )
-    };
 }
 
-macro_rules! sampler2 {
-    ($self:ident) => {
-        match $self.fractal {
-            Fractal::None => match $self.noise {
-                Noise::CellDistanceSq => make!($self, CellDistanceSq.jitter($self.jitter)),
-                Noise::CellDistance => make!($self, CellDistance.jitter($self.jitter)),
-                Noise::CellValue => make!($self, CellValue.jitter($self.jitter)),
-                Noise::OpenSimplex2 => make!($self, OpenSimplex2),
-                Noise::OpenSimplex2s => make!($self, OpenSimplex2s),
-                Noise::Perlin => make!($self, Perlin),
-                Noise::ValueCubic => make!($self, ValueCubic),
-                Noise::Value => make!($self, Value),
-            },
-            Fractal::Fbm => match $self.noise {
-                Noise::CellDistanceSq => make_fbm!($self, CellDistanceSq.jitter($self.jitter)),
-                Noise::CellDistance => make_fbm!($self, CellDistance.jitter($self.jitter)),
-                Noise::CellValue => make_fbm!($self, CellValue.jitter($self.jitter)),
-                Noise::OpenSimplex2 => make_fbm!($self, OpenSimplex2),
-                Noise::OpenSimplex2s => make_fbm!($self, OpenSimplex2s),
-                Noise::Perlin => make_fbm!($self, Perlin),
-                Noise::ValueCubic => make_fbm!($self, ValueCubic),
-                Noise::Value => make_fbm!($self, Value),
-            },
-            Fractal::Ridged => match $self.noise {
-                Noise::CellDistanceSq => make_ridged!($self, CellDistanceSq.jitter($self.jitter)),
-                Noise::CellDistance => make_ridged!($self, CellDistance.jitter($self.jitter)),
-                Noise::CellValue => make_ridged!($self, CellValue.jitter($self.jitter)),
-                Noise::OpenSimplex2 => make_ridged!($self, OpenSimplex2),
-                Noise::OpenSimplex2s => make_ridged!($self, OpenSimplex2s),
-                Noise::Perlin => make_ridged!($self, Perlin),
-                Noise::ValueCubic => make_ridged!($self, ValueCubic),
-                Noise::Value => make_ridged!($self, Value),
-            },
-            Fractal::PingPong => match $self.noise {
-                Noise::CellDistanceSq => make_ping_pong!($self, CellDistanceSq.jitter($self.jitter)),
-                Noise::CellDistance => make_ping_pong!($self, CellDistance.jitter($self.jitter)),
-                Noise::CellValue => make_ping_pong!($self, CellValue.jitter($self.jitter)),
-                Noise::OpenSimplex2 => make_ping_pong!($self, OpenSimplex2),
-                Noise::OpenSimplex2s => make_ping_pong!($self, OpenSimplex2s),
-                Noise::Perlin => make_ping_pong!($self, Perlin),
-                Noise::ValueCubic => make_ping_pong!($self, ValueCubic),
-                Noise::Value => make_ping_pong!($self, Value),
-            },
-        }
-    };
-}
-
-macro_rules! sampler3 {
+macro_rules! sampler {
     ($self:ident) => {
         match $self.fractal {
             Fractal::None => match $self.noise {
@@ -261,15 +187,13 @@ macro_rules! sampler3 {
                 Noise::CellValue => make!($self, CellValue.jitter($self.jitter)),
                 Noise::OpenSimplex2 => match $self.improve {
                     Improve::None => make!($self, OpenSimplex2),
-                    Improve::Default => make!($self, OpenSimplex2, Improve3),
-                    Improve::Xy => make!($self, OpenSimplex2, Improve3Xy),
-                    Improve::Xz => make!($self, OpenSimplex2, Improve3Xz),
+                    Improve::Xy => make!($self, OpenSimplex2.improve_xy()),
+                    Improve::Xz => make!($self, OpenSimplex2.improve_xz()),
                 },
                 Noise::OpenSimplex2s => match $self.improve {
                     Improve::None => make!($self, OpenSimplex2s),
-                    Improve::Default => make!($self, OpenSimplex2s, Improve3),
-                    Improve::Xy => make!($self, OpenSimplex2s, Improve3Xy),
-                    Improve::Xz => make!($self, OpenSimplex2s, Improve3Xz),
+                    Improve::Xy => make!($self, OpenSimplex2s.improve_xy()),
+                    Improve::Xz => make!($self, OpenSimplex2s.improve_xz()),
                 },
                 Noise::Perlin => make!($self, Perlin),
                 Noise::ValueCubic => make!($self, ValueCubic),
@@ -281,15 +205,13 @@ macro_rules! sampler3 {
                 Noise::CellValue => make_fbm!($self, CellValue.jitter($self.jitter)),
                 Noise::OpenSimplex2 => match $self.improve {
                     Improve::None => make_fbm!($self, OpenSimplex2),
-                    Improve::Default => make_fbm!($self, OpenSimplex2, Improve3),
-                    Improve::Xy => make_fbm!($self, OpenSimplex2, Improve3Xy),
-                    Improve::Xz => make_fbm!($self, OpenSimplex2, Improve3Xz),
+                    Improve::Xy => make_fbm!($self, OpenSimplex2.improve_xy()),
+                    Improve::Xz => make_fbm!($self, OpenSimplex2.improve_xz()),
                 },
                 Noise::OpenSimplex2s => match $self.improve {
                     Improve::None => make_fbm!($self, OpenSimplex2s),
-                    Improve::Default => make_fbm!($self, OpenSimplex2s, Improve3),
-                    Improve::Xy => make_fbm!($self, OpenSimplex2s, Improve3Xy),
-                    Improve::Xz => make_fbm!($self, OpenSimplex2s, Improve3Xz),
+                    Improve::Xy => make_fbm!($self, OpenSimplex2s.improve_xy()),
+                    Improve::Xz => make_fbm!($self, OpenSimplex2s.improve_xz()),
                 },
                 Noise::Perlin => make_fbm!($self, Perlin),
                 Noise::ValueCubic => make_fbm!($self, ValueCubic),
@@ -301,15 +223,13 @@ macro_rules! sampler3 {
                 Noise::CellValue => make_ridged!($self, CellValue.jitter($self.jitter)),
                 Noise::OpenSimplex2 => match $self.improve {
                     Improve::None => make_ridged!($self, OpenSimplex2),
-                    Improve::Default => make_ridged!($self, OpenSimplex2, Improve3),
-                    Improve::Xy => make_ridged!($self, OpenSimplex2, Improve3Xy),
-                    Improve::Xz => make_ridged!($self, OpenSimplex2, Improve3Xz),
+                    Improve::Xy => make_ridged!($self, OpenSimplex2.improve_xy()),
+                    Improve::Xz => make_ridged!($self, OpenSimplex2.improve_xz()),
                 },
                 Noise::OpenSimplex2s => match $self.improve {
                     Improve::None => make_ridged!($self, OpenSimplex2s),
-                    Improve::Default => make_ridged!($self, OpenSimplex2s, Improve3),
-                    Improve::Xy => make_ridged!($self, OpenSimplex2s, Improve3Xy),
-                    Improve::Xz => make_ridged!($self, OpenSimplex2s, Improve3Xz),
+                    Improve::Xy => make_ridged!($self, OpenSimplex2s.improve_xy()),
+                    Improve::Xz => make_ridged!($self, OpenSimplex2s.improve_xz()),
                 },
                 Noise::Perlin => make_ridged!($self, Perlin),
                 Noise::ValueCubic => make_ridged!($self, ValueCubic),
@@ -321,15 +241,13 @@ macro_rules! sampler3 {
                 Noise::CellValue => make_ping_pong!($self, CellValue.jitter($self.jitter)),
                 Noise::OpenSimplex2 => match $self.improve {
                     Improve::None => make_ping_pong!($self, OpenSimplex2),
-                    Improve::Default => make_ping_pong!($self, OpenSimplex2, Improve3),
-                    Improve::Xy => make_ping_pong!($self, OpenSimplex2, Improve3Xy),
-                    Improve::Xz => make_ping_pong!($self, OpenSimplex2, Improve3Xz),
+                    Improve::Xy => make_ping_pong!($self, OpenSimplex2.improve_xy()),
+                    Improve::Xz => make_ping_pong!($self, OpenSimplex2.improve_xz()),
                 },
                 Noise::OpenSimplex2s => match $self.improve {
                     Improve::None => make_ping_pong!($self, OpenSimplex2s),
-                    Improve::Default => make_ping_pong!($self, OpenSimplex2s, Improve3),
-                    Improve::Xy => make_ping_pong!($self, OpenSimplex2s, Improve3Xy),
-                    Improve::Xz => make_ping_pong!($self, OpenSimplex2s, Improve3Xz),
+                    Improve::Xy => make_ping_pong!($self, OpenSimplex2s.improve_xy()),
+                    Improve::Xz => make_ping_pong!($self, OpenSimplex2s.improve_xz()),
                 },
                 Noise::Perlin => make_ping_pong!($self, Perlin),
                 Noise::ValueCubic => make_ping_pong!($self, ValueCubic),
@@ -341,20 +259,20 @@ macro_rules! sampler3 {
 
 impl Config {
     pub fn sampler2(&self) -> Box<dyn Sample<2>> {
-        sampler2!(self)
+        sampler!(self)
     }
 
     #[cfg(feature = "nightly-simd")]
     pub fn sampler2a(&self) -> Box<dyn Sample<2, f32x2>> {
-        sampler2!(self)
+        sampler!(self)
     }
 
     pub fn sampler3(&self) -> Box<dyn Sample<3>> {
-        sampler3!(self)
+        sampler!(self)
     }
 
     #[cfg(feature = "nightly-simd")]
     pub fn sampler3a(&self) -> Box<dyn Sample<3, f32x4>> {
-        sampler3!(self)
+        sampler!(self)
     }
 }