+++ crates/noise-functions-config/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - **breaking:** update to `noise-functions` version 0.2.0
+- **breaking:** removed `Improve::Default`, `None` now behaves like `Default`
 
 ## 0.3.1 (2024-05-03)
 - **fix:** broken release with wrong dependency version